requirements.txt
black==22.8.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pycparser==2.21
PyGithub==1.55
PyJWT==2.5.0
PyNaCl==1.5.0
PyYAML==6.0
requests==2.28.1
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
